# oka_template/ooka_task/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API Documentation ",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "darius@matulionis.lt"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/ArsipKeluar": {
            "get": {
                "tags": [
                    "ArsipKeluar"
                ],
                "summary": "Get all Data arsipKeluar",
                "description": "Untuk menambah data arsip harus ada data surat keluar",
                "operationId": "index.arsipKeluar",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status values that needed to be considered for filter",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "default": "available"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        },
...
        "/api/ArsipKeluar/": {
            "get": {
                "tags": [
                    "ArsipKeluar"
                ],
                "summary": "Get 1 Data arsipKeluar",
                "description": "tambah /id",
                "operationId": "edit.arsipKeluar",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "default": "id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        },
...
        "/api/SuratKeluar": {
            "get": {
                "tags": [
                    "SuratKeluar"
                ],
                "summary": "Get all Data suratKeluar",
                "description": "",
                "operationId": "index.suratKeluar",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status values that needed to be considered for filter",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "default": "available"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        },
        "/api/SuratKeluar/store": {
            "post": {
                "tags": [
                    "SuratKeluar"
                ],
                "summary": "Tambah Data suratKeluar",
                "description": "",
                "operationId": "store.suratKeluar",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "default": "available"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        },
        "/api/SuratKeluar/{id}": {
            "get": {
                "tags": [
                    "SuratKeluar"
                ],
                "summary": "Show Data suratKeluar",
                "description": "",
                "operationId": "edit.suratKeluar",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status values that needed to be considered for filter",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "default": "available"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        },
        "/api/SuratKeluar/update": {
            "put": {
                "tags": [
                    "SuratKeluar"
                ],
                "summary": "update Data suratKeluar",
                "description": "kirim id lewat form",
                "operationId": "update.suratKeluar",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "default": "available"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        },
        "/api/SuratKeluar/destroy": {
            "delete": {
                "tags": [
                    "SuratKeluar"
                ],
                "summary": "hapus Data suratKeluar",
                "description": "kirim id lewat form",
                "operationId": "hapus.suratKeluar",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "default": "available"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        },
...
    }
}